+++ filter_variable.h
@@ -67,10 +67,10 @@ public:
 	}
 	virtual void update(void);
 private:
-	void update_fixed(const int16_t *in,
-		int16_t *lp, int16_t *bp, int16_t *hp);
-	void update_variable(const int16_t *in, const int16_t *ctl,
-		int16_t *lp, int16_t *bp, int16_t *hp);
+	void update_fixed(const int32_t *in,
+		int32_t *lp, int32_t *bp, int32_t *hp);
+	void update_variable(const int32_t *in, const int32_t *ctl,
+		int32_t *lp, int32_t *bp, int32_t *hp);
 	int32_t setting_fcenter;
 	int32_t setting_fmult;
 	int32_t setting_octavemult;
